src/requirements.txt
requests
ses-mailer-3